api/requirements.txt
opencv-python==4.6.0.66
imutils==0.5.4
numpy==1.23.4
kivy==2.1.0
python-dotenv==0.21.0
boto3==1.25.1
fastapi==0.85.1
uvicorn===0.18.3
pydantic==1.10.2
gunicorn==19.9.0
kivymd==1.1.1